move core perf check logic out of cli.py
